Добавили кастомные "ячейки в таблицу". Фотографии подгружаются из mock данных. В каждой ячейке отображается текущая дата. Лайк проставлен у четных ячеек.
