+++ ImageFeed/Helpers/UIColor+Extensions.swift
@@ -7,12 +7,12 @@
 
 import UIKit
 
-extension UIColor {
-        static var ypBlue: UIColor { UIColor(named: "YP Blue") ?? UIColor.green }
-        static var ypRed: UIColor { UIColor(named: "YP Red") ?? UIColor.red }
-        static var ypBlack: UIColor { UIColor(named: "YP Black") ?? UIColor.black}
-        static var ypBackground: UIColor { UIColor(named: "YP Background") ?? UIColor.darkGray }
-        static var ypGray: UIColor { UIColor(named: "YP Gray") ?? UIColor.gray }
-        static var ypWhite: UIColor { UIColor(named: "YP White") ?? UIColor.white}
-        static var ypWhiteAlpha50: UIColor { UIColor(named: "YP White (Alpha50)") ?? UIColor.white}
-     }
+//extension UIColor {
+//        static var ypBlue: UIColor { UIColor(named: "YP Blue") ?? UIColor.green }
+//        static var ypRed: UIColor { UIColor(named: "YP Red") ?? UIColor.red }
+//        static var ypBlack: UIColor { UIColor(named: "YP Black") ?? UIColor.black}
+//        static var ypBackground: UIColor { UIColor(named: "YP Background") ?? UIColor.darkGray }
+//        static var ypGray: UIColor { UIColor(named: "YP Gray") ?? UIColor.gray }
+//        static var ypWhite: UIColor { UIColor(named: "YP White") ?? UIColor.white}
+//        static var ypWhiteAlpha50: UIColor { UIColor(named: "YP White (Alpha50)") ?? UIColor.white}
+//     }
