--- ImageFeed/ImagesList/ImagesListViewController.swift.swift
@@ -1,58 +1,0 @@
-//
-//  ViewController.swift
-//  ImageFeed
-//
-//  Created by Ринат Шарафутдинов on 01.08.2023.
-//
-
-import UIKit
-
-class ImagesListViewController: UIViewController {
-    
-    @IBOutlet private var tableView: UITableView!
-    
-    private lazy var dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .long
-        formatter.timeStyle = .none
-        return formatter
-    }()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-    
-
-
-
-}
-
-extension ImagesListViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
-    }
-}
-
-extension ImagesListViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
-    }
-    
-    func tableView( _ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: ImagesListCell.reuseIdentifier, for: indexPath)
-        
-        guard let imageListCell = cell as? ImagesListCell else {
-            return UITableViewCell()
-        }
-        
-        configCell(for: imageListCell)
-        return imageListCell
-    }
-    
-    func configCell(for cell: ImagesListCell) {
-    }
-    
-    
-}
-
